feat: 리뷰 반영 - css 수정
